refactor: split headless.py into providers/wrappers/headless + lazy model sync
Split the monolithic headless.py into three focused modules:
- providers.py: AgentProvider dataclass, AGENT_PROVIDERS registry, lookup
- wrappers.py: shell wrapper generation for all agent CLI commands
- headless.py: headless-only command construction and config resolution

Rename HeadlessProvider → AgentProvider and HEADLESS_PROVIDERS →
AGENT_PROVIDERS since these describe all modes, not just headless.

Move vibe-model-sync from bashrc (ran on every login, ~500ms Python
startup) to the vibe() wrapper with a pure-bash mtime guard — sync
only runs when the user actually invokes vibe and the cache is stale.
